+++ src/main/java/rbasamoyai/createbigcannons/cannon_control/cannon_mount/CannonMountInterfaceBlockEntity.java
@@ -9,9 +9,11 @@ import com.simibubi.create.content.kinetics.transmission.sequencer.SequencerInst
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import rbasamoyai.createbigcannons.cannon_control.contraption.PitchOrientedContraptionEntity;
 
 public abstract class CannonMountInterfaceBlockEntity extends KineticBlockEntity {
 
@@ -35,12 +37,18 @@ public abstract class CannonMountInterfaceBlockEntity extends KineticBlockEntity
 		super.onSpeedChanged(previousSpeed);
 		this.sequencedAngleLimit = -1;
 
-		if (this.sequenceContext != null && this.sequenceContext.instruction() == SequencerInstructions.TURN_ANGLE) {
+        PitchOrientedContraptionEntity mountedContraption = this.parent.mountedContraption;
+        if (mountedContraption != null && Math.signum(previousSpeed) != Math.signum(this.getSpeed()) && previousSpeed != 0) {
+            if (!mountedContraption.isStalled())
+                this.applyRotationAndAngle();
+        }
+
+        if (this.sequenceContext != null && this.sequenceContext.instruction() == SequencerInstructions.TURN_ANGLE) {
 			this.sequencedAngleLimit = this.sequenceContext.getEffectiveValue(getTheoreticalSpeed()) * 0.125f;
 		}
 	}
 
-	public void tryUpdateSpeed() {
+    public void tryUpdateSpeed() {
 		if (this.preventSpeedUpdate > 0)
 			return;
 		this.warnOfMovement();
@@ -54,7 +62,22 @@ public abstract class CannonMountInterfaceBlockEntity extends KineticBlockEntity
 	@Override public void sendData() { this.parent.sendData(); }
 	@Override public void setChanged() { this.parent.setChanged(); }
 
-	public static class PitchInterface extends CannonMountInterfaceBlockEntity {
+    protected abstract void applyRotationAndAngle();
+
+    @Override
+    protected void write(CompoundTag tag, boolean clientPacket) {
+        super.write(tag, clientPacket);
+        if (this.sequencedAngleLimit >= 0)
+            tag.putDouble("SequencedAngleLimit", this.sequencedAngleLimit);
+    }
+
+    @Override
+    protected void read(CompoundTag tag, boolean clientPacket) {
+        super.read(tag, clientPacket);
+        this.sequencedAngleLimit = tag.contains("SequencedAngleLimit") ? tag.getDouble("SequencedAngleLimit") : -1;
+    }
+
+    public static class PitchInterface extends CannonMountInterfaceBlockEntity {
 		public PitchInterface(BlockEntityType<?> typeIn, BlockPos pos, BlockState state, CannonMountBlockEntity parent) {
 			super(typeIn, pos, state, parent);
 		}
@@ -67,6 +90,12 @@ public abstract class CannonMountInterfaceBlockEntity extends KineticBlockEntity
             li.add(this.worldPosition.offset(pos1));
             li.add(this.worldPosition.subtract(pos1));
             return li;
+        }
+
+        @Override
+        protected void applyRotationAndAngle() {
+            this.parent.setPitch(Math.round(this.parent.cannonPitch));
+            this.parent.applyRotation();
         }
     }
 
@@ -81,6 +110,12 @@ public abstract class CannonMountInterfaceBlockEntity extends KineticBlockEntity
             List<BlockPos> li = new ArrayList<>();
             li.add(this.worldPosition.relative(vertical));
             return li;
+        }
+
+        @Override
+        protected void applyRotationAndAngle() {
+            this.parent.setYaw(Math.round(this.parent.cannonYaw));
+            this.parent.applyRotation();
         }
     }